Intake Changes 17:37 Fixed Field Centric, optimized current limit in Intake, and removed PitchIntake systems.

+++ src/main/java/frc/robot/commands/ExtendIntakeCommand.java
@@ -22,7 +22,7 @@ public class ExtendIntakeCommand extends CommandBase{
      * @param voltage Gets voltage to determine speed and retract or Extend
      * @param extendIntakeMotorSubsystem Extend intake motor subsystem
      */
-    public ExtendIntakeCommand(int voltage, ExtendIntakeMotorSubsystem extendIntakeMotorSubsystem){
+    public ExtendIntakeCommand(double voltage, ExtendIntakeMotorSubsystem extendIntakeMotorSubsystem){
         m_voltage = voltage;
         m_ExtendIntakeMotorSubsystem = extendIntakeMotorSubsystem;
         addRequirements(m_ExtendIntakeMotorSubsystem);
